Smarter Guards
Guards are now able to follow a route by placing waypoints, if the player hits the "bar" of vision and raycast hits player, the guards will stop following waypoints and go to players position.

Known Problems - Guards a bit slow on rotation in this small map, and don't play well together in tight spaces.

Will be looking to add a better cone of vision to guards, and more quality of life improvements on the waypoints