Minor guard improvments
Added some functionality to guards to allow them to rotate and not move. and updated levels to use it when needed.
Updated simulated modal used in pathing to help prevent guards from hitting the walls as much.

Minor tweaks to guards vision distance  and field of view as they didn't feel too fun at those distances.